requirements.txt
streamlit
langchain
langchain-community
pypdf
PyMuPDF
google-generativeai
python-dotenv
pillow
python-docx
openai
deepgram-sdk